requirements.txt
-i https://pypi.org/simple
aiofiles==23.2.1 ; python_version >= '3.7'
blinker==1.7.0 ; python_version >= '3.8'
click==8.1.7 ; python_version >= '3.7'
exceptiongroup==1.2.0 ; python_version >= '3.7'
flask==3.0.0 ; python_version >= '3.8'
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0 ; python_full_version >= '3.6.1'
hpack==4.0.0 ; python_full_version >= '3.6.1'
hypercorn==0.16.0 ; python_version >= '3.8'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
importlib-metadata==7.0.1 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jsonify==0.5
markupsafe==2.1.3 ; python_version >= '3.7'
priority==2.0.0 ; python_full_version >= '3.6.1'
pyjwt==2.8.0
quart==0.19.4
taskgroup==0.0.0a4 ; python_version < '3.11'
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.9.0 ; python_version < '3.10'
werkzeug==3.0.1 ; python_version >= '3.8'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
zipp==3.17.0 ; python_version >= '3.8'
